Update Program.cs
i've added function that in the case when the older category has no record that the younger has the program insert results from younger to older category, i've also solved some problems with erros that i haven't been able to reslove without using try catch. Mainly it is incorporated in TextSanitizer function. I've  got rid of URLexists function becouse it turned out that it doesn't serve any purpose i mean it wasn't functioning as i was expecting it to